Improve toplevel decisions
The NO_FUEL_REGUGEE event is an "evt-"  event that can be a starting
event when you wait at the sector.  Previously it was being printed
both nested and as a toplevel event, with the text only appearing on
the toplevel event.

Now, it only prints the nested version to make it clearer that it
belongs to the NO_FUEL group. Furthermore, this nested version prints
the event text, so you can ctrl-f it.